pyproject.toml
[project]
name = "parallel-appim"
version = "0.1.0"
description = "Appium server gateway hub for parallel testing on multiple devices"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "appium-python-client>=5.2.4,<6.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.4.0",
    "httpx>=0.25.0",
    "pytest>=7.4.0",
    "pytest-xdist>=3.3.0",
    "pytest-asyncio>=0.21.0"
]

[project.scripts]
appium-hub = "appium_hub.main:main"

[project.optional-dependencies]
dev = [
    "pytest-html>=3.1.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
